+++ peernode/src/peer-node/producer/mod.rs
@@ -9,44 +9,6 @@ use std::sync::Arc;
 
 use anyhow::Result;
 
-pub async fn run(market: String) -> Result<()> {
-    let mut client = MarketClient::new(market).await?;
-
-    // Load the files
-    let file_map = Arc::new(files::FileMap::default());
-    file_map.add_dir("files/**/*", 100).await?;
-
-    // Launch the HTTP server in the background
-    let http_file_map = file_map.clone();
-    tokio::spawn(async move {
-        if let Err(e) = http::run(http_file_map, String::from("8080")).await {
-            eprintln!("HTTP server error: {}", e);
-        }
-    });
-
-    // Register the files with the market service
-    let hash = file_map.get_hashes().await;
-    for hash in hash {
-        // TODO: Find a way to get the public IP address
-        println!("Producer: Registering file with hash {}", hash);
-        client
-            .register_file(
-                "id".to_string(),
-                "name".to_string(),
-                "127.0.0.1".to_string(),
-                8080,
-                100,
-                hash,
-            )
-            .await?;
-    }
-
-    // Never return
-    tokio::signal::ctrl_c().await?;
-
-    Ok(())
-}
-
 pub async fn start_server(
     files: HashMap<String, PathBuf>,
     prices: HashMap<String, i64>,
